Fixed minor warnings in Windows.

+++ src/debugger/DebuggerParser.cxx
@@ -108,7 +108,7 @@ string DebuggerParser::run(const string& command)
   getArgs(command, verb);
   commandResult.str("");
 
-  for(size_t i = 0; i < commands.size(); ++i)
+  for(int i = 0; i < commands.size(); ++i)
   {
     if(BSPF::equalsIgnoreCase(verb, commands[i].cmdString))
     {
